Implements validation for frontend Change the directory of initial data Add Exception Handler class

+++ spring_project/src/main/java/com/example/spring_project/application/service/UserService.java
@@ -1,10 +1,12 @@
 package com.example.spring_project.application.service;
 
 import java.util.List;
+import java.util.Objects;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.example.spring_project.domain.errors.UserNotFoundException;
 import com.example.spring_project.domain.model.user.NewUser;
 import com.example.spring_project.domain.model.user.SearchUser;
 import com.example.spring_project.domain.model.user.User;
@@ -24,15 +26,27 @@ public class UserService {
     }
 
     public User findById (Integer id) {
-        return repository.findById(id); 
+        User user = repository.findById(id);
+        if (Objects.isNull(user)) {
+            throw new UserNotFoundException(id);
+        }
+        return user;
     }
-    public void update (User user) {
+        
+    public User update (User user) {
+        if (Objects.isNull(repository.findById(user.id()))) {
+            throw new UserNotFoundException(user.id());
+        }
         repository.update(user);
+        return repository.findById(user.id());
     }
     public void add (NewUser user) {
         repository.add(user);
     }
     public void delete (Integer id) {
+        if (Objects.isNull(repository.findById(id))) {
+            throw new UserNotFoundException(id);
+        }
         repository.delete(id);
     }
